Declare game function
Declared game function which uses a for loop to invoke playRound 5 times
. Every win increments the score variable by one. If playRound returns
"Tie" it will decrement the loop counter by one to simulate replaying
the round and disregarding the round result. If you lose the round, the
loop counter will increment by one but score will not. This simulates a
"best of 5" game mode.

Additionally, in the case of a tie, the playRound function will no
longer invoke itself but, rather, this will be accounted for in the game
function as described above.