Print individual round results
Declared function printRoundResult to log individual round results to
the console. This function takes the player choice and round result as
arguments and uses an if .. else conditional statement to evaluate
the result. If the result is a win or loss, it then accounts for the
player choice using a switch statement to determine what exact message
to print i.e. rock beats paper. If the result is a tie the function
returns and nothing is printed.

printRoundResult is then invoked in every loop of the game function to
print the result of each round.