Declare printGameResult
Declared printGameResult which takes the result of the game function as
its argument and prints this number to the console along with some text.
This function is invoked within the game function after the for loop is
completed. Whereas before, the game function printed the results
directly.

Also changed other function expressions from let to const since these
values won't change.