Prompt user for playerChoice
Instead of hard coding the value of playerChoice, we prompt the user
to input rock, paper or scissors with the getPlayerChoice function.
This function is invoked at the beginning of the game function to get
the initial playerChoice and within the for loop everytime a round is
decided. If the round results in a tie, the function is not invoked but
instead the round is replayed with the same playerChoice until a winner
is decided.