Display ties
Changed the code so that upon a tie, a message will be displayed to the
user telling them the round was a tie and prompting them to chose again
to redo the round. This is repeated until a winner is achieved.

This makes it feel more like an authentic game of rock, paper, scissors.
Whereas before, upon a tie, the tie was not displayed to the user and
the computer simply selected again until a winner was achieved.